requirements.txt
torch==1.2.0
matplotlib
numpy
gpytorch
mnist
absl-py